+++ git-config/src/file/access/raw.rs
@@ -3,7 +3,7 @@ use std::{borrow::Cow, collections::HashMap};
 use bstr::BStr;
 
 use crate::{
-    file::{mutable::value::EntryData, Index, MutableMultiValue, MutableSection, MutableValue, Size},
+    file::{mutable::multi_value::EntryData, Index, MutableMultiValue, MutableSection, MutableValue, Size},
     lookup,
     parse::{section, Event},
     File,
