+++ git-config/src/file/mod.rs
@@ -10,8 +10,9 @@ use bstr::BStr;
 mod mutable;
 
 pub use mutable::{
+    multi_value::MutableMultiValue,
     section::{MutableSection, SectionBody, SectionBodyIter},
-    value::{MutableMultiValue, MutableValue},
+    value::MutableValue,
 };
 
 mod init;
