+++ git-config/tests/file/access/mod.rs
@@ -1,4 +1,3 @@
 mod mutate;
 mod raw;
 mod read_only;
-mod write;
