+++ git-config/tests/file/mod.rs
@@ -30,3 +30,4 @@ mod access;
 mod impls;
 mod init;
 mod mutable;
+mod write;
